+++ Finance/Presentation/Auth/View/AuthInput.swift
@@ -1,0 +1,12 @@
+//
+//  AuthInput.swift
+//  Finance
+//
+//  Created by Александр Меренков on 22.06.2023.
+//
+
+import Foundation
+
+protocol AuthInput: AnyObject {
+    func showAlert(message: String)
+}
