+++ Finance/Presentation/Profile/Presenter/ProfilePresenterOutput.swift
@@ -1,0 +1,12 @@
+//
+//  ProfilePresenterOutput.swift
+//  Finance
+//
+//  Created by Александр Меренков on 22.06.2023.
+//
+
+import Foundation
+
+protocol ProfilePresenterOutput: AnyObject {
+    
+}
