+++ Finance/Presentation/Profile/View/ProfileViewController.swift
@@ -30,7 +30,7 @@ final class ProfileViewController: UIViewController {
         label.textAlignment = .center
         return label
     }()
-    private let authController = AuthController()
+    private let authController = AuthCoordinator(authAssembly: AuthAssembly()).start()
 
     private lazy var currentCurrencyBtn = Utils().menuItemButton(image: "currencies.png", title: "Текущая валюта")
     private lazy var logOutBtn = Utils().menuItemButton(image: "logout.png", title: "Выход")
@@ -130,7 +130,7 @@ final class ProfileViewController: UIViewController {
 
 extension ProfileViewController: ProfileInput {
     func showAuth() {
-        authController.delegate = self
+//        authController.delegate = self
         self.view.addSubview(authController.view)
         self.addChild(authController)
         authController.didMove(toParent: self)
